fix: refactor datalayer's test file
