+++ components/forms/Form/Form.test.js
@@ -2,9 +2,9 @@ import React from "react";
 import { cleanup, render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import Form from "./Form";
-import mockedDataLayer from "../../../lib/integration/common";
+import mockedDataLayer from "../../../lib/integration/helpers";
 
-jest.mock("../../../lib/integration/common", () => ({
+jest.mock("../../../lib/integration/helpers", () => ({
   submitToAPI: jest.fn(() => {}),
 }));
