+++ lib/tests/markdown.test.js
@@ -1,6 +1,6 @@
 import markdown from "../markdown";
 
-jest.mock("../integration/common", () => ({
+jest.mock("../integration/helpers", () => ({
   extractFormData: jest.fn(() => ["one", "two", "three"]),
   rehydrateFormResponses: jest.fn(() => null),
 }));
