+++ pages/api/submit.tsx
@@ -3,8 +3,8 @@ import type { NextApiRequest, NextApiResponse } from "next";
 import { LambdaClient, InvokeCommand } from "@aws-sdk/client-lambda";
 import formidable, { Fields, Files } from "formidable";
 import convertMessage from "@lib/markdown";
-import { getFormByID, rehydrateFormResponses } from "@lib/integration/common";
-import { getSubmissionByID } from "@lib/integration/common";
+import { getFormByID, rehydrateFormResponses } from "@lib/integration/helpers";
+import { getSubmissionByID } from "@lib/integration/helpers";
 import { logMessage } from "@lib/logger";
 import { PublicFormSchemaProperties, Responses } from "@lib/types";
 import { checkOne } from "@lib/flags";
